requirements.txt
absl-py==2.1.0
aiofiles==24.1.0
aiohappyeyeballs==2.3.0
aiohttp>=3.8.0
aioice==0.9.0
aiortc==1.10.1
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyascii==0.3.2
anyio==3.7.1
asgiref==3.8.1
asttokens==3.0.0
attrs==23.1.0
audioread==3.0.1
av==13.1.0
babel==2.17.0
backoff==2.2.1
bangla==0.0.2
bcrypt==4.2.1
beautifulsoup4==4.13.3
black==25.1.0
blinker==1.9.0
blis==1.2.0
bnnumerizer==0.0.2
bnunicodenormalizer==0.1.7
Brotli @ file:///croot/brotli-split_1736182456865/work
build==1.2.2.post1
cachetools==5.5.1
catalogue==2.0.10
certifi @ file:///croot/certifi_1738623731865/work/certifi
cffi @ file:///home/conda/feedstock_root/build_artifacts/cffi_1725560564262/work
chardet==5.2.0
charset-normalizer @ file:///croot/charset-normalizer_1721748349566/work
chroma-hnswlib==0.7.6
chromadb>=0.4.0
chromadb==0.6.3
click==8.1.8
cloudpathlib==0.20.0
cmake==3.31.4
coloredlogs==15.0.1
colorlog==6.9.0
confection==0.1.5
contourpy==1.3.1
coqpit==0.0.17
coverage==7.6.12
cryptography==44.0.0
cycler==0.12.1
cymem==2.0.11
Cython==3.0.11
dataclasses-json==0.6.7
dateparser==1.1.8
decorator==5.1.1
defusedxml==0.7.1
deprecated==1.2.18
deprecation==2.1.0
distro==1.9.0
dnspython==2.7.0
docopt==0.6.2
durationpy==0.9
einops==0.8.1
elevenlabs==1.50.7
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
encodec==0.1.1
executing==2.2.0
faiss-cpu==1.10.0
fastapi==0.115.8
filelock @ file:///croot/filelock_1700591183607/work
flake8==7.1.1
flask==3.1.0
flatbuffers==25.1.24
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2024.6.1
g2pkk==0.1.2
gitdb==4.0.12
GitPython==3.1.44
gmpy2 @ file:///croot/gmpy2_1738085463648/work
google-api-core==2.24.1
google-api-python-client==2.160.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-crc32c==1.6.0
googleapis-common-protos==1.66.0
gotrue==2.11.3
greenlet==3.1.1
grpcio==1.70.0
gruut==2.2.3
gruut-ipa==0.13.0
gruut_lang_de==2.0.1
gruut_lang_en==2.0.1
gruut_lang_es==2.0.1
gruut_lang_fr==2.0.2
gTTS==2.5.4
h11==0.14.0
h2==4.2.0
hangul-romanize==0.1.0
hpack==4.1.0
httpcore==1.0.7
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.28.1
humanfriendly==10.0
hyperframe==6.1.0
idna @ file:///croot/idna_1714398848350/work
ifaddr==0.2.0
importlib_metadata==8.5.0
importlib_resources==6.5.2
inflect==7.5.0
iniconfig==2.0.0
ipython==8.32.0
itsdangerous==2.2.0
jamo==0.4.1
jedi==0.19.2
jieba==0.42.1
Jinja2 @ file:///croot/jinja2_1737760107953/work
jiter==0.8.2
joblib==1.4.2
jsonlines==1.2.0
jsonpatch==1.33
jsonpickle==4.0.1
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
keybert==0.9.0
kiwisolver==1.4.8
kubernetes==32.0.0
langchain==0.3.18
langchain-community==0.0.20
langchain-core==0.3.34
langchain-text-splitters==0.3.6
langcodes==3.5.0
langsmith==0.3.8
language_data==1.3.0
lazy_loader==0.4
librosa>=0.10.0
lit==18.1.8
llama-index==0.9.26
llvmlite==0.44.0
marisa-trie==1.2.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe @ file:///croot/markupsafe_1738584038848/work
marshmallow==3.26.1
matplotlib==3.10.0
matplotlib-inline==0.1.7
mccabe==0.7.0
mdurl==0.1.2
mmh3==5.1.0
monotonic==1.6
more-itertools==10.6.0
mpmath @ file:///croot/mpmath_1690848262763/work
msgpack==1.1.0
multidict==6.1.0
murmurhash==1.0.12
mypy-extensions==1.0.0
narwhals==1.26.0
neo4j>=5.10.0
nest-asyncio==1.6.0
networkx==2.8.8
nltk==3.9.1
num2words==0.5.14
numba==0.61.0
numpy>=1.24.0
nvidia-cublas-cu11==11.10.3.66
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu11==8.5.0.96
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu11==10.2.10.91
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu11==11.7.4.91
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu11==2.14.3
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu11==11.7.91
nvidia-nvtx-cu12==12.4.127
oauthlib==3.2.2
onnxruntime==1.20.1
openai==0.28.1
openai-whisper==20240930
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
opentelemetry-api==1.30.0
opentelemetry-exporter-otlp-proto-common==1.30.0
opentelemetry-exporter-otlp-proto-grpc==1.30.0
opentelemetry-instrumentation==0.51b0
opentelemetry-instrumentation-asgi==0.51b0
opentelemetry-instrumentation-fastapi==0.51b0
opentelemetry-proto==1.30.0
opentelemetry-sdk==1.30.0
opentelemetry-semantic-conventions==0.51b0
opentelemetry-util-http==0.51b0
orjson==3.10.15
overrides==7.7.0
packaging==23.2
pandas==1.5.3
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
Pillow>=10.0.0
pillow @ file:///croot/pillow_1738010226202/work
platformdirs==4.3.6
pluggy==1.5.0
pocketsphinx==5.0.4
pooch==1.8.2
postgrest==0.19.3
posthog==3.11.0
preshed==3.0.9
prompt_toolkit==3.0.50
propcache==0.2.1
proto-plus==1.26.0
protobuf==5.29.3
psutil==6.1.1
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==19.0.0
pyasn1==0.6.1
pyasn1-modules==0.4.1
PyAudio==0.2.14
pycodestyle==2.12.1
pycparser @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_pycparser_1733195786/work
pydantic>=2.0.0
pydantic-settings==2.7.1
pydantic_core==2.27.2
pydeck==0.9.1
pyee==12.1.1
pyflakes==3.2.0
pygame==2.6.1
Pygments==2.19.1
PyJWT==2.10.1
pylibsrtp==0.11.0
pynndescent==0.5.13
pyOpenSSL==25.0.0
pyparsing==3.2.1
PyPika==0.48.9
pypinyin==0.53.0
pyproject-hooks==1.2.0
pysbd==0.3.4
PySocks @ file:///work/ci_py311/pysocks_1676822712504/work
pytest==7.4.0
pytest-asyncio>=0.21.0
pytest-cov==4.1.0
pytest-mock==3.14.0
python-crfsuite==0.9.11
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytube==15.0.0
pytz==2025.1
pyvis==0.3.2
PyYAML @ file:///croot/pyyaml_1728657952215/work
realtime==2.3.0
referencing==0.36.2
regex==2024.11.6
reportlab==4.3.0
requests @ file:///croot/requests_1730999120400/work
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
resampy==0.4.3
rich==13.9.4
rpds-py==0.22.3
rsa==4.9
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.15.1
sentence-transformers==3.4.1
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
smmap==5.0.2
sniffio==1.3.0
sounddevice==0.5.1
soundfile==0.13.1
soupsieve==2.6
soxr==0.5.0.post1
spacy==3.8.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SpeechRecognition==3.14.1
SQLAlchemy==2.0.38
srsly==2.5.1
st-supabase-connection==2.0.1
stack-data==0.6.3
starlette==0.45.3
storage3==0.11.3
streamlit==1.42.0
streamlit-webrtc==0.47.9
StrEnum==0.4.15
SudachiDict-core==20250129
SudachiPy==0.6.10
supabase==2.13.0
supafunc==0.9.3
sympy==1.13.1
tenacity==8.2.2
tensorboard==2.18.0
tensorboard-data-server==0.7.2
thinc==8.3.4
threadpoolctl==3.5.0
tiktoken==0.8.0
timm==1.0.14
tokenizers==0.21.0
toml==0.10.2
torch>=2.0.0
torchaudio==2.6.0
torchvision==0.21.0
tornado==6.4.2
tqdm==4.67.1
trainer==0.0.36
traitlets==5.14.3
transformers>=4.30.0
triton==3.2.0
TTS==0.22.0
typeguard==2.13.3
typer==0.15.1
typing-inspect==0.9.0
typing_extensions @ file:///croot/typing_extensions_1734714854207/work
tzdata==2025.1
tzlocal==5.2
umap-learn==0.5.7
Unidecode==1.3.8
uritemplate==4.1.1
urllib3 @ file:///croot/urllib3_1737133630106/work
uvicorn==0.34.0
uvloop==0.21.0
vidgear==0.3.3
wasabi==1.1.3
watchdog==6.0.0
watchfiles==1.0.4
wcwidth==0.2.13
weasel==0.4.1
websocket-client==1.8.0
websockets==14.2
Werkzeug==3.1.3
wikipedia==1.4.0
wrapt==1.17.2
yarl==1.18.3
youtube-transcript-api==0.6.3
yt-dlp==2025.1.26
zipp==3.21.0
zstandard==0.23.0
